requirements.txt
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0
Pillow>=8.0.0
numpy>=1.19.0 